Disallow socket files removal for Windows systems
Former-commit-id: 3b5bd598e59d0dee61e0ec3ccda94705843531cb
Former-commit-id: 616a0656878347854d6296956e6e3125d0cbd5a0

+++ cmd/skywire-visor/commands/root.go
@@ -13,6 +13,7 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"runtime"
 	"syscall"
 	"time"
 
@@ -193,7 +194,7 @@ func (rc *runConf) runVisor() *runConf {
 
 	time.Sleep(startDelay)
 
-	if rc.conf.Dmsgpty != nil {
+	if rc.conf.Dmsgpty != nil && runtime.GOOS != "windows" {
 		if err := visor.UnlinkSocketFiles(rc.conf.Dmsgpty.CLIAddr); err != nil {
 			rc.logger.Fatal("failed to unlink socket files: ", err)
 		}
